fix call size default behavior for librarians and indices dtype for vectorizer

+++ src/sciterra/vectorization/sbert.py
@@ -71,6 +71,6 @@ class SBERTVectorizer(Vectorizer):
         # We don't have to deal with OOV, so we always return full list of ids
         return {
             "embeddings": np.array(embeddings),
-            "success_indices": np.arange(len(embeddings)),
+            "success_indices": np.arange(len(embeddings), dtype=int),
             "fail_indices": np.array([], dtype=int),
         }
